@@ -26,18 +26,17 @@ classifiers = [
 
 [tool.poetry.dependencies]
 aiohttp = "^3.9.1"
-defusedxml = "^0.7.1"
-python = "^3.11"
+python = "^3.12"
 
 [tool.poetry.group.dev.dependencies]
 bandit = "^1.7.6"
 mypy = "^1.11.0"
-poethepoet = "^0.27.0"
-ruff = "^0.5.0"
+poethepoet = "^0.28.0"
+ruff = "^0.6.3"
 
 [tool.poetry.group.test.dependencies]
 pytest = "^8.0.0"
-pytest-asyncio = "^0.23.3"
+pytest-asyncio = "^0.24.0"
 pytest-cov = "^5.0.0"
 pytest-dotenv = "^0.5.2"
 
